updates: sheet Integrations,styling
