+++ src/Pages/InterestGroups/Dropdown.module.css
@@ -1,3 +1,7 @@
+.container{
+    overflow: hidden;
+    border-radius: 10px;
+}
 .table{
     display: flex;
     justify-content: center;
@@ -11,12 +15,20 @@
     background-color: rgb(246, 132, 44);
     padding: 10px 20px;
     color: #fefefe;
+    cursor: zoom-out;
 }
-.container{
-    border-radius: 15px;
-    margin:10px 0;
+.open{
+    border-radius: 10px;
+    cursor: zoom-in;
+}
 
+.scrollContainer{
+    border-radius: 10px;
+
+    overflow-x: auto;
+    margin:10px 0;
 }
+
 .first{
     display: flex;
     align-items: center;
@@ -31,12 +43,22 @@
 }
 .tabler{
     width: 100%;
+    border-collapse: collapse;
 }
 .box{
     padding:10px 0; 
 }
 .block{
     text-align: center;
+    padding: 8px;
+}
+.block:nth-child(1),.block:nth-child(4),.block:nth-child(5){
+    min-width: 20px;
+    padding:0 35px;
+}
+.block:nth-child(2),.block:nth-child(3){
+    min-width: 50px;
+    min-width: max(200px,5vw);
 }
 .head,.block{
     font-size: 13.5px;
@@ -48,8 +70,27 @@
 .linked{
     font-weight: 400;
     text-decoration: underline;
+    cursor: pointer;
 }
 .block:nth-child(2),.head:nth-child(2){
     text-align: left;
+    width: 200px;
 }
 
+.modal{
+    position: fixed;
+    top:0;
+    left:0;
+    z-index: 99999;
+    background-color: rgba(0, 0, 0, 0.534);
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+}
+.iframe{
+    aspect-ratio: 16/9;
+    width: 80%;
+
+}
