+++ src/Pages/InterestGroups/IComponents/Board/Board.jsx
@@ -1,0 +1,34 @@
+import styles from "../../InterestGroup.module.css"
+import fvimg from '../../assets/fvimg.png'
+const InterestGroupBoard = ({ children, igSummary }) => {
+    return (<div className={styles.main_container}>
+        <div className={styles.first_view_container}>
+            <div className={styles.first_view}>
+                <div className={styles.fv_texts}>
+                    {igSummary &&
+                        <>
+                            <p className={styles.fv_heading}>
+                                {igSummary?.name}
+                            </p>
+                            <p className={styles.fv_content}>
+                                {igSummary?.description}
+                            </p>
+                            {igSummary?.office_hour && <p className={styles.officehrs}>
+                                <span> Office Hours: </span> {igSummary?.office_hour + ' @ ' + igSummary?.offlice_place}
+                            </p>}
+                        </>
+                    }
+                </div>
+                <div className={styles.fv_image}>
+                    <img
+                        src={fvimg}
+                        alt="Group Learning pic"
+                        className={styles.fv_img}
+                    />
+                </div>
+            </div>
+        </div>
+        {children}
+    </div>)
+}
+export default InterestGroupBoard
