+++ src/Pages/LearningCircles/Home/Home.jsx
@@ -26,6 +26,7 @@ import InterestCard from "../../../Components/InterestCard/InterestCard";
 import Preloader from "../../../Components/Preloader/Preloader";
 import { useRef } from "react";
 import SheetAPI from "../../../Utils/SheetAPI";
+import getInterestGroupsData from "../../InterestGroups/Utils/getInterestGroupsData";
 
 const Home = ({
   backenderr,
@@ -76,7 +77,7 @@ const Home = ({
           setMembers(response.data.data.members);
         })
         .catch(function (error) {
-          // console.log(error);
+          console.log(error);
         });
     }
   }, [code]);
@@ -109,19 +110,11 @@ const Home = ({
         });
     }
   }, [college, interest]);
-  const firstCall = useRef(true)
   const [summary, setSummary] = useState([])
+
   useEffect(() => {
-    if (firstCall.current) {
-      SheetAPI(
-        "https://docs.google.com/spreadsheets/d/1F3rdjrJ8cS12oR90ClqNNrw10WgWYOQU0A6G4WHdRbA/edit#gid=1291474495",
-        "Summary",
-        setSummary
-      )
-    }
-  }, [])
-  useEffect(() => {
-    console.log(summary)
+    if (!summary.length)
+      getInterestGroupsData(summary, undefined, setSummary)
   }, [summary])
   if (districts) {
     return (
@@ -426,7 +419,7 @@ const Home = ({
             {summary.map((ig, index) => (
               <InterestCard
                 key={index}
-                id={ig?.page_link}
+                id={ig?.id}
                 interestgroup={ig?.name}
                 interestgroupdescription={
                   ig.description
